+++ app/templates/layout.html
@@ -5,12 +5,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
 	<style>
 		* {box-sizing: border-box;}
-		
-		body { 
+
+		body {
 		  margin: 0;
 		  font-family: Arial, Helvetica, sans-serif;
 		}
-		
+
 		#navbar {
 		  overflow: hidden;
 		  background-color: #f1f1f1;
@@ -21,38 +21,38 @@
 		  top: 0;
 		  z-index: 99;
 		}
-		
+
 		#navbar a {
 		  float: left;
 		  color: black;
 		  text-align: center;
 		  padding: 12px;
 		  text-decoration: none;
-		  font-size: 18px; 
+		  font-size: 18px;
 		  line-height: 25px;
 		  border-radius: 4px;
 		}
-		
+
 		#navbar #logo {
 		  font-size: 35px;
 		  font-weight: bold;
 		  transition: 0.4s;
 		}
-		
+
 		#navbar a:hover {
 		  background-color: #ddd;
 		  color: black;
 		}
-		
+
 		#navbar a.active {
 		  background-color: dodgerblue;
 		  color: white;
 		}
-		
+
 		#navbar-right {
 		  float: right;
 		}
-		
+
 		@media screen and (max-width: 580px) {
 		  #navbar {
 			padding: 20px 10px !important;
@@ -75,16 +75,16 @@
 		<div id="navbar-right">
 		  <a class="active" href="#home">Home</a>
 		  <a href="#contact">Contact</a>
-		  <a href="signin.html>Sign In</a>
+		  <a href="{{  url_for('loginHandler.login') }}">Sign In</a>
 		</div>
 	  </div>
-	
+
 	<div id="content" style="margin-top:210px;padding:15px 15px 2500px;font-size:30px">{% block content %}{% endblock %}</div>
 
 	<script>
 		// When the user scrolls down 80px from the top of the document, resize the navbar's padding and the logo's font size
 		window.onscroll = function() {scrollFunction()};
-		
+
 		function scrollFunction() {
 		  if (document.body.scrollTop > 80 || document.documentElement.scrollTop > 80) {
 			document.getElementById("navbar").style.padding = "30px 10px";
