+++ app/templates/login.html
@@ -9,5 +9,3 @@
     <input type="submit" value="Log In">
   </form>
 {% endblock %}
-
-{% endblock %}
