+++ requirements.txt
@@ -1,2 +1,3 @@
+wtforms
 Flask==1.1.1
 flask-login==0.5.0
